requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.7.4.post0; python_version >= '3.6'
appdirs==1.4.4
apscheduler==3.6.3
asttokens==2.0.5
async-generator==1.10; python_version >= '3.5'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
automat==20.2.0
beautifulsoup4==4.9.3
better-exceptions==0.3.3
blinker==1.4
brotli==1.0.9
certifi==2022.6.15
cffi==1.15.0
cfscrape==2.1.1
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cloudscraper==1.2.60
colorama==0.4.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
constantly==15.1.0
cryptography==37.0.2; python_version >= '3.6'
cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
decorator==5.1.1; python_version >= '3.5'
discord.py==1.7.3
discord==1.7.3
distlib==0.3.4
executing==0.8.3
filelock==3.7.1; python_version >= '3.7'
get-mac==0.8.3
h11==0.13.0; python_version >= '3.6'
h2==4.1.0; python_version >= '3.6'
hpack==4.0.0; python_full_version >= '3.6.1'
hyperframe==6.0.1; python_version >= '3.6'
hyperlink==21.0.0
icecream==2.1.2
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==4.12.0; python_version >= '3.7'
incremental==21.3.0
itemadapter==0.6.0; python_version >= '3.6'
itemloaders==1.0.4; python_version >= '3.6'
jmespath==1.0.1; python_version >= '3.7'
kaitaistruct==0.9
lxml==4.9.0
manga-py==1.33.2
multidict==6.0.2; python_version >= '3.7'
outcome==1.2.0; python_version >= '3.7'
packaging==21.3; python_version >= '3.6'
parsel==1.6.0
patool==1.12
pdf2image==1.14.0
pillow==9.1.1; python_version >= '3.7'
pipenv==2020.11.15
progressbar2==4.0.0; python_version >= '3.7'
protego==0.2.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pure-python-adb==0.3.0.dev0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycryptodome==3.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydispatcher==2.0.5; platform_python_implementation == 'CPython'
pyee==8.2.2
pyexecjs==1.5.1
pygments==2.12.0; python_version >= '3.6'
pyopenssl==22.0.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyppeteer==1.0.2
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-telegram-bot==13.1
python-utils==3.3.3; python_version >= '3.7'
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2022.1
pyyaml==5.4
queuelib==1.6.2; python_version >= '3.5'
requests-file==1.5.1
requests-toolbelt==0.9.1
requests==2.25.1
scrapy==2.6.1
selenium-wire==4.6.4
selenium==4.3.0
service-identity==21.1.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sortedcontainers==2.4.0
soupsieve==2.3.2.post1; python_version >= '3.0'
sqlalchemy==1.3.22
tldextract==3.3.0; python_version >= '3.7'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tornado==6.1; python_version >= '3.5'
tqdm==4.64.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
trio-websocket==0.9.2; python_version >= '3.5'
trio==0.21.0; python_version >= '3.7'
twisted==22.4.0; python_full_version >= '3.6.7'
typing-extensions==4.2.0; python_version >= '3.7'
tzdata==2022.1; python_version >= '3.6'
tzlocal==4.2; python_version >= '3.6'
urllib3[secure,socks]==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
virtualenv-clone==0.5.4
virtualenv==20.2.2
w3lib==1.22.0
websockets==10.3; python_version >= '3.7'
werkzeug==2.0.3; python_version >= '3.6'
wsproto==1.1.0; python_version >= '3.7'
xmltodict==0.12.0
yarl==1.7.2; python_version >= '3.6'
zipp==3.4.0
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zstandard==0.18.0; python_version >= '3.6'
